[refactor] LangChain 도입을 위한 프로젝트 구조 개편
