+++ src/api/v2/endpoint/chat.py
@@ -1,0 +1,2 @@
+# src/api/v2/endpoint/chat.py
+# LangChain 기반 API 엔드포인트 (TODO: 구현 예정)
