+++ src/v2_services/chat_service.py
@@ -1,0 +1,2 @@
+# src/services_langchain/chat_service.py
+# LangChain 기반 RAG 서비스 (TODO: 구현 예정)
