+++ src/v2_services/compare_service.py
@@ -1,0 +1,2 @@
+# src/services_langchain/compare_service.py
+# LangChain 기반 법률 비교 서비스 (TODO: 구현 예정)
